[feat]#70 로그아웃 기능 구현에 따라 FCM 토큰 저장 로직 Application -> LoginActivity로 이전
